certtool: avoid 1-byte write buffer overrun when parsing template
Previously, when parsing a template file with a number of key value
pairs, certtool could write a NUL byte after the heap buffer, causing
a memory corruption. This fixes the issue by allocating the NUL byte.
Reported by David Aitel.

+++ tests/cert-tests/template-test.sh
@@ -297,6 +297,19 @@ if test "${rc}" != "0"; then
 	exit ${rc}
 fi
 
+# Test generation with too many other names
+
+"${CERTTOOL}" --generate-request \
+	--load-privkey "${srcdir}/data/template-test.key" \
+	--template "${srcdir}/templates/template-too-many-othernames.tmpl" \
+	--outfile ${TMPFILE} 2>/dev/null
+rc=$?
+
+if test "${rc}" != "0"; then
+	echo "Test with too many othernames failed"
+	exit ${rc}
+fi
+
 rm -f ${TMPFILE}
 
 exit 0
